+++ app/api/business/contacts/resource.py
@@ -3,8 +3,8 @@ import uuid
 
 from flask import request
 from flask_restplus import Resource
-from src.app.api.restplus import api
-from src.app.api.business.contacts.models import contact
+from app.api.restplus import api
+from app.api.business.contacts.models import contact
 
 log = logging.getLogger(__name__)
