+++ requirements.txt
@@ -1,2 +1,2 @@
 -r tasks/requirements.txt
--r src/app/requirements.txt
+-r app/requirements.txt
