--- src/api.swagger.yml
@@ -1,69 +1,0 @@
-openapi: "3.0.0"
-info:
-  version: 1.0.0
-  title: SKP CRM api
-  license:
-    name: MIT
-servers:
-  - url: tbd
-paths:
-  /contacts/{contactId}:
-    get:
-      parameters:
-      - in: path
-        name: contactId
-        schema:
-          type: integer
-        required: true
-      summary: retrieve a contact
-      operationId: getContact
-      tags:
-        - Contact Management
-      responses:
-        '200':
-          description: A Contact
-          content:
-            application/json:    
-              schema:
-                $ref: "#/components/schemas/Contact"
-        '404':
-          $ref: '#/components/responses/NotFound'
-        default:
-          description: unexpected error
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Error"
-components:
-  responses:
-    NotFound:
-      description: The specified resource was not found
-      content:
-        application/json:
-          schema:
-            $ref: '#/components/schemas/Error'
-  schemas:
-    Contact:
-      properties:
-        id:
-          type: string
-          format: uuid
-        surname:
-          type: string
-        givennames:
-          type: object
-          properties:
-            firstName:
-              type: string
-            middleName:
-              type: string
-    Error:
-      required:
-        - code
-        - message
-      properties:
-        code:
-          type: integer
-          format: int32
-        message:
-          type: string
